requirements.txt
# Core dependencies
streamlit>=1.28.0
folium>=0.14.0
streamlit-folium>=0.15.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Geospatial libraries (install these after fixing compatibility)
geopandas>=0.14.0
shapely>=2.0.0
geopy>=2.4.0

# Additional utilities
pytz>=2024.1

# Supabase for historical data logging
supabase>=2.0.0
